模板可能存在的慢查询

+++ src/backend/ci/core/process/api-process/src/main/kotlin/com/tencent/devops/process/engine/common/Timeout.kt
@@ -43,7 +43,7 @@ object Timeout {
 
     val MAX_MINUTES = TimeUnit.DAYS.toMinutes(MAX_JOB_RUN_DAYS).toInt() // 7 * 24 * 60 = 10080 分钟 = 最多超时7天
 
-    val CONTAINER_MAX_MILLS = TimeUnit.MINUTES.toMillis(MAX_MINUTES.toLong()).toInt() + 1 // 毫秒+1
+    val CONTAINER_MAX_MILLS = TimeUnit.MINUTES.toMillis(MAX_MINUTES.toLong()) + 1 // 毫秒+1
 
     fun transMinuteTimeoutToMills(timeoutMinutes: Int?): Pair<Int, Long> {
         var minute = timeoutMinutes ?: DEFAULT_TIMEOUT_MIN
